Only report supported hvac_modes and fan_modes
This updates climate.py to report the capabilities reported by the Smartthinq model information.

Our LP1419IVSM only supports `cool`, `fan_only` and `dry` `hvac_modes`, and the presence of `auto` would cause hass-smartthinq to set the temperature to the minimum whenever it's activated.

This updates it to filter the `hvac_modes` to only those which are listed in the device's SupportOpMode list, and to filter `fan_modes` to only those listed in the device's SupportWindStrength list.